pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-limiter>=3.10.1",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.3",
    "openai>=1.65.1",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.38",
    "werkzeug>=3.1.3",
    "pypdf2>=3.0.1",
    "faiss-cpu>=1.10.0",
]
